+++ cla/inc/bplib_cla.h
@@ -189,53 +189,53 @@ BPLib_Status_t BPLib_CLA_ContactsTblValidateFunc(void *TblData);
 /**
  * \brief     Find the requested contact ID in the Contacts Configuration and pass that information to the
  *            CLA proxy to configure the CLA with
- * \param[in] ContactId (uint16_t) Contact ID from the Contacts Configuration to setup
+ * \param[in] ContactId (uint32_t) Contact ID from the Contacts Configuration to setup
  * \return    Execution status
  * \retval    BPLIB_SUCCESS: Successful execution
  * \retval    BPLIB_CLA_UNKNOWN_CONTACT: Contact ID could not be found in the Contacts Configuration
  * \retval    BPLIB_CLA_CONTACTS_MAX_REACHED: Setting up contact would exceed maximum simultaneous
  *                                            contacts allowed
  */
-BPLib_Status_t BPLib_CLA_ContactSetup(uint16_t ContactId);
+BPLib_Status_t BPLib_CLA_ContactSetup(uint32_t ContactId);
 
 /**
  * \brief     Pass Contact ID to start, on to CLA proxy
- * \param[in] ContactId (uint16_t) Contact ID from the Contacts Configuration to start
+ * \param[in] ContactId (uint32_t) Contact ID from the Contacts Configuration to start
  * \return    Execution status
  * \retval    BPLIB_SUCCESS: Successful execution
  * \retval    BPLIB_CLA_ERROR: TODO
  */
-BPLib_Status_t BPLib_CLA_ContactStart(uint16_t ContactId);
+BPLib_Status_t BPLib_CLA_ContactStart(uint32_t ContactId);
 
 /**
  * \brief     Pass Contact ID to stop, on to CLA proxy
- * \param[in] ContactId (uint16_t) Contact ID from the Contacts Configuration to stop
+ * \param[in] ContactId (uint32_t) Contact ID from the Contacts Configuration to stop
  * \return    Execution status
  * \retval    BPLIB_SUCCESS: Successful execution
  * \retval    BPLIB_CLA_ERROR: TODO
  */
-BPLib_Status_t BPLib_CLA_ContactStop(uint16_t ContactId);
+BPLib_Status_t BPLib_CLA_ContactStop(uint32_t ContactId);
 
 /**
  * \brief     If the contact has been stopped, deconfigure the CLA via BI, CT, EBP, and CLA
- * \param[in] ContactId (uint16_t) Contact ID from the Contacts Configuration to teardown
+ * \param[in] ContactId (uint32_t) Contact ID from the Contacts Configuration to teardown
  * \return    Execution status
  * \retval    BPLIB_SUCCESS: Successful execution
  * \retval    BPLIB_CLA_CONTACT_RUNNING: Contact has not been stopped
  */
-BPLib_Status_t BPLib_CLA_ContactTeardown(uint16_t ContactId);
+BPLib_Status_t BPLib_CLA_ContactTeardown(uint32_t ContactId);
 
 /**
   * \brief     Get access to the run state of a particular contact
-  * \param[in] ContactId (uint16_t) Contact ID from the Contacts Configuration of whose run
+  * \param[in] ContactId (uint32_t) Contact ID from the Contacts Configuration of whose run
   *                                 state should be returned
   * \return    Contact run state
   */
-BPLib_CLA_ContactRunState_t BPLib_CLA_GetContactRunState(uint16_t ContactId);
+BPLib_CLA_ContactRunState_t BPLib_CLA_GetContactRunState(uint32_t ContactId);
 
 /**
   * \brief     Set the run state of the provided contact
-  * \param[in] ContactId (uint16_t) Contact ID from the Contacts Configuration whose run state is being requested to change
+  * \param[in] ContactId (uint32_t) Contact ID from the Contacts Configuration whose run state is being requested to change
   * \param[in] RunState  (BPLib_CLA_ContactRunState_t) Requested run state of the provided contact ID from the Contacts Configuration
   * \return    Execution status
   * \retval    BPLIB_SUCCESS: Successfully changed the run state of the provided contact ID to the provided run state
@@ -243,6 +243,6 @@ BPLib_CLA_ContactRunState_t BPLib_CLA_GetContactRunState(uint16_t ContactId);
   * \retval    BPLIB_CLA_INCORRECT_STATE: The intended run state for the contact with the provided contact ID was incompatible with the
   *                                       current run state of the contact
   */
- BPLib_Status_t BPLib_CLA_SetContactRunState(uint16_t ContactId, BPLib_CLA_ContactRunState_t RunState);
+ BPLib_Status_t BPLib_CLA_SetContactRunState(uint32_t ContactId, BPLib_CLA_ContactRunState_t RunState);
 
 #endif /* BPLIB_CLA_H */
